Module rsx_shared::traits

Traits

FromPrimitive
A generic trait for converting a number to a value.

NumCast
An interface for casting between machine scalars.

TAnyPartialEq

TCallback

TClientPosition

TClientRect

TClientSize

TComputedStyles

TDOMNode

TDOMText

TDOMTree

TDimensionsInfo

TEncodedFont

TEncodedImage

TEvent

TEventManager

TFileCache

TFontCache

TFontInstanceKey

TFontKey

TFontKeysAPI

TGenericEvent

TGlyphInstance

TGlyphStore

TImageCache

TImageKeysAPI

TInheritedStyles

TKeyboardEvent

TLayoutNode

TMeasuredImage

TMediaKey

TMemoryAPI

TMemoryAddrAPI

TMouseEvent

TProp

TResourceGroup

TRunner

TRuntime

TShapedText

TStyleDeclarations

TUIEvent

ToPrimitive
A generic trait for converting a value to a number.